Addressed all comments in PR, refactored muygps_torch test code
